change array uint16 to int

+++ Inc/ui/scope.h
@@ -5,7 +5,7 @@
 #define ADC_bufsize 320
 typedef struct
 {
-	uint16_t buf[2][ADC_bufsize];
+	int buf[2][ADC_bufsize];
 	float max;
 	float min;
 	float p2p;
